requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15,<3.16
drf-spectacular>=0.27,<0.28
djangorestframework-simplejwt>=5.3,<5.4
django-filter>=24.3,<25
psycopg2-binary>=2.9,<3
celery>=5.3,<5.4
redis>=5.0,<5.1
python-dateutil>=2.8,<2.9
pydantic>=2.6,<2.8
pydantic-settings>=2.2,<2.3
structlog>=24.1,<25
pillow>=10.1,<10.2
boto3>=1.34,<1.35
phonenumbers>=8.13,<8.14
django-cors-headers>=4.3,<4.4
dj-database-url>=2.1,<2.2
fastapi>=0.110,<0.111
uvicorn[standard]>=0.29,<0.30
PyJWT>=2.8,<2.9
gunicorn>=21.2,<21.3
opensearch-py>=2.5,<2.6
